subo un cambio

+++ main.cpp
@@ -207,7 +207,7 @@ vector<vector<uint16_t>> datosAMatriz(uchar &datos, int ancho, int alto) {
 
 
 int main(int argc, char * argv[]) {
-    vector<vector<short int>>* matriz;
+    vector<vector<double>>* matriz;
     matriz = leerCSV("dicom_csv2/1.2.826.0.1.3680043.2.656.1.138.1.csv");
 
     cout << (*matriz)[0][0] << endl;
